refactor(links): split label text extraction
Move link-label clipping and text normalization into a dedicated module so link annotation extraction can focus on targets, boxes, and destination resolution.
